webfontloader partial renamed to match export function name
